wgpu_types

Module backend

Backend, Backends, and backend-specific options.

Structs

BackendOptions

Options that are passed to a given backend.

Backends

Represents the backends that wgpu will use.

Dx12BackendOptions

Configuration for the DX12 backend.

GlBackendOptions

Configuration for the OpenGL/OpenGLES backend.

NoopBackendOptions

Configuration for the noop backend.

Enums

Backend

Backends supported by wgpu.

Dx12Compiler

Selects which DX12 shader compiler to use.

Dx12SwapchainKind

Selects which kind of swapchain to use on DX12.

Dx12UseFrameLatencyWaitableObject

Whether and how to use a waitable handle obtained from GetFrameLatencyWaitableObject.

DxcShaderModel

DXC shader model.

GlFenceBehavior

Dictate the behavior of fences in OpenGL.

Gles3MinorVersion

Selects which OpenGL ES 3 minor version to request.