wgpu_types

Module binding

Bind groups and the bindings in them.

Structs

BindGroupLayoutEntry

Describes a single binding inside a bind group.

ShaderStages

Describes the shader stages that a binding will be visible from.

ShaderStagesBits

Bits from ShaderStages in array form

ShaderStagesWGPU

Shader stages that are available in WGPU only when targeting native. For all usages, see ShaderStages.

ShaderStagesWebGPU

Shader stages that are defined in the WebGPU specification. For all usages, see ShaderStages.

Enums

BindingType

Type of a binding in a bind group layout.

BufferBindingType

Specific type of a buffer binding.

SamplerBindingType

Specific type of a sampler binding.