wgpu_types::binding

Struct ShaderStages

#[repr(C)]
pub struct ShaderStages {
    pub features_webgpu: ShaderStagesWebGPU,
    pub features_wgpu: ShaderStagesWGPU,
}

Describes the shader stages that a binding will be visible from.

These can be combined so something that is visible from both vertex and fragment shaders can be defined as:

ShaderStages::VERTEX | ShaderStages::FRAGMENT

Corresponds to WebGPU GPUShaderStageFlags.

Fields
features_webgpu: ShaderStagesWebGPUfeatures_wgpu: ShaderStagesWGPU
Implementations

impl ShaderStages

pub(crate) const FLAGS: &'static [Flag<Self>]

pub const NONE: Self

Binding is not visible from any shader stage.

pub const VERTEX: Self

Binding is visible from the vertex shader of a render pipeline.

pub const FRAGMENT: Self

Binding is visible from the fragment shader of a render pipeline.

pub const COMPUTE: Self

Binding is visible from the compute shader of a compute pipeline.

pub const VERTEX_FRAGMENT: Self

Binding is visible from the vertex and fragment shaders of a render pipeline.

pub const TASK: Self

Binding is visible from the task shader of a mesh pipeline.

pub const MESH: Self

Binding is visible from the mesh shader of a mesh pipeline.

pub const RAY_GENERATION: Self

Binding is visible from the ray generation shader of a ray tracing pipeline.

pub const ANY_HIT: Self

Binding is visible from the ray any hit shader of a ray tracing pipeline.

pub const CLOSEST_HIT: Self

Binding is visible from the ray closest hit shader of a ray tracing pipeline.

pub const MISS: Self

Binding is visible from the ray miss shader of a ray tracing pipeline.

pub const fn bits(&self) -> ShaderStagesBits

Gets the set flags as a container holding an array of bits.

pub const fn empty() -> Self

Returns self with no flags set.

pub const fn all() -> Self

Returns self with all flags set.

pub const fn contains(self, other: Self) -> bool

Whether all the bits set in other are all set in self

pub const fn intersects(self, other: Self) -> bool

Returns whether any bit set in self matched any bit set in other.

pub const fn is_empty(self) -> bool

Returns whether there is no flag set.

pub const fn is_all(self) -> bool

Returns whether the struct has all flags set.

pub const fn union(self, other: Self) -> Self

Bitwise or - self | other

pub const fn intersection(self, other: Self) -> Self

Bitwise and - self & other

pub const fn difference(self, other: Self) -> Self

Bitwise and of the complement of other - self & !other

pub const fn symmetric_difference(self, other: Self) -> Self

Bitwise xor - self ^ other

pub const fn complement(self) -> Self

Bitwise not - !self

pub fn set(&mut self, other: Self, set: bool)

Calls Self::insert if set is true and otherwise calls Self::remove.

pub fn insert(&mut self, other: Self)

Inserts specified flag(s) into self

pub fn remove(&mut self, other: Self)

Removes specified flag(s) from self

pub fn toggle(&mut self, other: Self)

Toggles specified flag(s) in self

pub const fn from_bits(bits: ShaderStagesBits) -> Option<Self>

Takes in ShaderStagesBits and returns None if there are invalid bits or otherwise Self with those bits set

pub const fn from_bits_truncate(bits: ShaderStagesBits) -> Self

Takes in ShaderStagesBits and returns Self with only valid bits (all other bits removed)

pub const fn from_bits_retain(bits: ShaderStagesBits) -> Self

Takes in ShaderStagesBits and returns Self with all bits that were set without removing invalid bits

pub fn from_name(name: &str) -> Option<Self>

Takes in a bitflags flag name (in SCREAMING_SNAKE_CASE) and returns Self if it matches or none if the name does not match the name of any of the flags. Name is capitalisation dependent.

[impl FromStr] can be used to recognize kebab-case names, like are used in the WebGPU spec.

pub fn from_internal_flags( features_webgpu: ShaderStagesWebGPU, features_wgpu: ShaderStagesWGPU, ) -> Self

Combines the flags from the internal flags into the entire flags struct

pub const fn iter(&self) -> Iter<ShaderStages>

Returns an iterator over the set flags.

pub const fn iter_names(&self) -> IterNames<ShaderStages>

Returns an iterator over the set flags and their names.

These are bitflags names in SCREAMING_SNAKE_CASE.