wgpu_types
Enum BlendFactor

#[repr(C)]
pub enum BlendFactor {

    Zero = 0,
    One = 1,
    Src = 2,
    OneMinusSrc = 3,
    SrcAlpha = 4,
    OneMinusSrcAlpha = 5,
    Dst = 6,
    OneMinusDst = 7,
    DstAlpha = 8,
    OneMinusDstAlpha = 9,
    SrcAlphaSaturated = 10,
    Constant = 11,
    OneMinusConstant = 12,
    Src1 = 13,
    OneMinusSrc1 = 14,
    Src1Alpha = 15,
    OneMinusSrc1Alpha = 16,

}

Alpha blend factor.

Corresponds to WebGPU GPUBlendFactor. Values using Src1 require Features::DUAL_SOURCE_BLENDING and can only be used with the first render target.

For further details on how the blend factors are applied, see the analogous functionality in OpenGL: https://www.khronos.org/opengl/wiki/Blending#Blending_Parameters.

Variants

Zero = 0

0.0

One = 1

1.0

Src = 2

S.component

OneMinusSrc = 3

1.0 - S.component

SrcAlpha = 4

S.alpha

OneMinusSrcAlpha = 5

1.0 - S.alpha

Dst = 6

D.component

OneMinusDst = 7

1.0 - D.component

DstAlpha = 8

D.alpha

OneMinusDstAlpha = 9

1.0 - D.alpha

SrcAlphaSaturated = 10

min(S.alpha, 1.0 - D.alpha)

Constant = 11

Constant

OneMinusConstant = 12

1.0 - Constant

Src1 = 13

S1.component

OneMinusSrc1 = 14

1.0 - S1.component

Src1Alpha = 15

S1.alpha

OneMinusSrc1Alpha = 16

1.0 - S1.alpha

Implementations

impl BlendFactor

pub fn ref_second_blend_source(&self) -> bool

Returns true if the blend factor references the second blend source.

Note that the usage of those blend factors require Features::DUAL_SOURCE_BLENDING.