wgpu_types

Module instance

Types for dealing with Instances

Structs

BackendOptions

Options that are passed to a given backend.

Dx12BackendOptions

Configuration for the DX12 backend.

GlBackendOptions

Configuration for the OpenGL/OpenGLES backend.

InstanceDescriptor

Options for creating an instance.

InstanceFlags

Instance debugging flags.

MemoryBudgetThresholds

Memory budget thresholds used by backends to try to avoid high memory pressure situations.

NoopBackendOptions

Configuration for the noop backend.

Enums

Dx12Compiler

Selects which DX12 shader compiler to use.

Dx12UseFrameLatencyWaitableObject

Whether and how to use a waitable handle obtained from GetFrameLatencyWaitableObject.

DxcShaderModel

DXC shader model.

GlFenceBehavior

Dictate the behavior of fences in OpenGL.

Gles3MinorVersion

Selects which OpenGL ES 3 minor version to request.