wgpu_types

Module limits

Limits and downlevel-related types.

Macros

with_limits 🔒

Invoke a macro for each of the limits.

Structs

DownlevelCapabilities

Lists various ways the underlying platform does not conform to the WebGPU standard.

DownlevelFlags

Binary flags listing features that may or may not be present on downlevel adapters.

DownlevelLimits

Represents the sets of additional limits on an adapter, which take place when running on downlevel backends.

Limits

Represents the sets of limits an adapter/device supports.

Enums

ShaderModel

Collections of shader features a device supports if they support less than WebGPU normally allows.