wgpu_types::macros

Trait ConstDefaultHelper

pub(crate) trait ConstDefaultHelper {
    const DEFAULT: Self;
}

Default value of a type, as a const item.

This trait is not public and is used solely to help ConstDefault provide inherent const fn default()s.

Specifically, it is implemented both for wgpu-types types and std types, so that the macro ConstDefault can call it without needing to understand the type.

Required Associated Constants

const DEFAULT: Self

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.