wgpu_types

Module render

Types for configuring render passes and render pipelines (except for vertex attributes).

Structs

BlendComponent

Describes a blend component of a BlendState.

BlendState

Describe the blend state of a render pipeline, within ColorTargetState.

ColorTargetState

Describes the color state of a render pipeline.

ColorWrites

Color write mask. Disabled color channels will not be written to.

DepthBiasState

Describes the biasing setting for the depth target.

DepthStencilState

Describes the depth/stencil state in a render pipeline.

DispatchIndirectArgs

Argument buffer layout for dispatch_indirect commands.

DrawIndexedIndirectArgs

Argument buffer layout for draw_indexed_indirect commands.

DrawIndirectArgs

Argument buffer layout for draw_indirect commands.

MultisampleState

Describes the multi-sampling state of a render pipeline.

Operations

Pair of load and store operations for an attachment aspect.

PrimitiveState

Describes the state of primitive assembly and rasterization in a render pipeline.

RenderBundleDepthStencil

Describes the depth/stencil attachment for render bundles.

RenderBundleDescriptor

Describes a RenderBundle.

StencilFaceState

Describes stencil state in a render pipeline.

StencilState

State of the stencil operation (fixed-pipeline stage).

Enums

BlendFactor

Alpha blend factor.

BlendOperation

Alpha blend operation.

CompareFunction

Comparison function used for depth and stencil operations.

Face

Face of a vertex.

FrontFace

Vertex winding order which classifies the “front” face of a triangle.

IndexFormat

Format of indices used with pipeline.

LoadOp

Operation to perform to the output attachment at the start of a render pass.

PolygonMode

Type of drawing mode for polygons

PrimitiveTopology

Primitive type the input mesh is composed of.

StencilOperation

Operation to perform on the stencil value.

StoreOp

Operation to perform to the output attachment at the end of a render pass.