wgpu_types

Struct BufferTextureCopyInfo

pub struct BufferTextureCopyInfo {

    pub copy_width: u64,
    pub copy_height: u64,
    pub depth_or_array_layers: u64,
    pub offset: u64,
    pub block_size_bytes: u64,
    pub block_width_texels: u64,
    pub block_height_texels: u64,
    pub width_blocks: u64,
    pub height_blocks: u64,
    pub row_bytes_dense: u64,
    pub row_stride_bytes: u64,
    pub image_stride_rows: u64,
    pub image_stride_bytes: u64,
    pub image_rows_dense: u64,
    pub image_bytes_dense: u64,
    pub bytes_in_copy: u64,

}

Information about a copy between a buffer and a texture.

Mostly used for internal calculations, but useful nonetheless. Generated by TexelCopyBufferLayout::get_buffer_texture_copy_info.

Fields
copy_width: u64

The width of the copy region in pixels.

copy_height: u64

The height of the copy region in pixels.

depth_or_array_layers: u64

The depth of the copy region in pixels.

offset: u64

The offset in the buffer where the copy starts.

block_size_bytes: u64

The size of a single texture texel block in bytes.

block_width_texels: u64

The number of texel in a texel block in the x direction.

block_height_texels: u64

The number of texel in a texel block in the y direction.

width_blocks: u64

The width of the copy region in blocks.

height_blocks: u64

The height of the copy region in blocks.

row_bytes_dense: u64

The number of bytes in the last row of the copy region.

row_stride_bytes: u64

The stride in bytes between the start of one row in an image and the next row in the same image.

This includes any padding between one row and the next row.

image_stride_rows: u64

The stride in rows between the start of one image and the next image.

image_stride_bytes: u64

The stride in bytes between the start of one image and the next image.

image_rows_dense: u64

The number of rows in a densely packed list of images.

This is the number of rows in the image that are actually used for texel data, and does not include any padding rows, unlike image_stride_rows.

image_bytes_dense: u64

The number of bytes in a densely packed list of images.

This is the number of bytes in the image that are actually used for texel data, or are used for padding between rows. Padding at the end of the last row and between images is not included.

bytes_in_copy: u64

The total number of bytes in the copy region.

This includes all padding except the padding after the last row in the copy.