wgpu_types

Struct BufferUsages

#[repr(C)]
pub struct BufferUsages {
    pub buffer_usages_webgpu: BufferUsagesWebGPU,
    pub buffer_usages_wgpu: BufferUsagesWGPU,
}

Different ways that you can use a buffer.

The usages determine what kind of memory the buffer is allocated from and what actions the buffer can partake in.

Specifying only usages the application will actually perform may increase performance. Additionally, on the WebGL backend, there are restrictions on BufferUsages::INDEX; see DownlevelFlags::UNRESTRICTED_INDEX_BUFFER for more information.

Corresponds to WebGPU GPUBufferUsageFlags.

Fields
buffer_usages_webgpu: BufferUsagesWebGPUbuffer_usages_wgpu: BufferUsagesWGPU
Implementations

impl BufferUsages

pub(crate) const FLAGS: &'static [Flag<Self>]

pub const MAP_READ: Self

Allow a buffer to be mapped for reading using Buffer::map_async + Buffer::get_mapped_range. This does not include creating a buffer with BufferDescriptor::mapped_at_creation set.

If Features::MAPPABLE_PRIMARY_BUFFERS isn’t enabled, the only other usage a buffer may have is COPY_DST.

pub const MAP_WRITE: Self

Allow a buffer to be mapped for writing using Buffer::map_async + Buffer::get_mapped_range_mut. This does not include creating a buffer with BufferDescriptor::mapped_at_creation set.

If Features::MAPPABLE_PRIMARY_BUFFERS feature isn’t enabled, the only other usage a buffer may have is COPY_SRC.

pub const COPY_SRC: Self

Allow a buffer to be the source buffer for a CommandEncoder::copy_buffer_to_buffer or CommandEncoder::copy_buffer_to_texture operation.

pub const COPY_DST: Self

Allow a buffer to be the destination buffer for a CommandEncoder::copy_buffer_to_buffer, CommandEncoder::copy_texture_to_buffer, CommandEncoder::clear_buffer or Queue::write_buffer operation.

pub const INDEX: Self

Allow a buffer to be the index buffer in a draw operation.

pub const VERTEX: Self

Allow a buffer to be the vertex buffer in a draw operation.

pub const UNIFORM: Self

Allow a buffer to be a BufferBindingType::Uniform inside a bind group.

pub const STORAGE: Self

Allow a buffer to be a BufferBindingType::Storage inside a bind group.

pub const INDIRECT: Self

Allow a buffer to be the indirect buffer in an indirect draw call.

pub const QUERY_RESOLVE: Self

Allow a buffer to be the destination buffer for a CommandEncoder::resolve_query_set operation.

pub const BLAS_INPUT: Self

Allows a buffer to be used as input for a bottom level acceleration structure build

pub const TLAS_INPUT: Self

Allows a buffer to be used as input for a top level acceleration structure build

pub const fn bits(&self) -> BufferUsagesBits

Gets the set flags as a container holding an array of bits.

pub const fn empty() -> Self

Returns self with no flags set.

pub const fn all() -> Self

Returns self with all flags set.

pub const fn contains(self, other: Self) -> bool

Whether all the bits set in other are all set in self

pub const fn intersects(self, other: Self) -> bool

Returns whether any bit set in self matched any bit set in other.

pub const fn is_empty(self) -> bool

Returns whether there is no flag set.

pub const fn is_all(self) -> bool

Returns whether the struct has all flags set.

pub const fn union(self, other: Self) -> Self

Bitwise or - self | other

pub const fn intersection(self, other: Self) -> Self

Bitwise and - self & other

pub const fn difference(self, other: Self) -> Self

Bitwise and of the complement of other - self & !other

pub const fn symmetric_difference(self, other: Self) -> Self

Bitwise xor - self ^ other

pub const fn complement(self) -> Self

Bitwise not - !self

pub fn set(&mut self, other: Self, set: bool)

Calls Self::insert if set is true and otherwise calls Self::remove.

pub fn insert(&mut self, other: Self)

Inserts specified flag(s) into self

pub fn remove(&mut self, other: Self)

Removes specified flag(s) from self

pub fn toggle(&mut self, other: Self)

Toggles specified flag(s) in self

pub const fn from_bits(bits: BufferUsagesBits) -> Option<Self>

Takes in BufferUsagesBits and returns None if there are invalid bits or otherwise Self with those bits set

pub const fn from_bits_truncate(bits: BufferUsagesBits) -> Self

Takes in BufferUsagesBits and returns Self with only valid bits (all other bits removed)

pub const fn from_bits_retain(bits: BufferUsagesBits) -> Self

Takes in BufferUsagesBits and returns Self with all bits that were set without removing invalid bits

pub fn from_name(name: &str) -> Option<Self>

Takes in a bitflags flag name (in SCREAMING_SNAKE_CASE) and returns Self if it matches or none if the name does not match the name of any of the flags. Name is capitalisation dependent.

[impl FromStr] can be used to recognize kebab-case names, like are used in the WebGPU spec.

pub fn from_internal_flags( buffer_usages_webgpu: BufferUsagesWebGPU, buffer_usages_wgpu: BufferUsagesWGPU, ) -> Self

Combines the flags from the internal flags into the entire flags struct

pub const fn iter(&self) -> Iter<BufferUsages>

Returns an iterator over the set flags.

pub const fn iter_names(&self) -> IterNames<BufferUsages>

Returns an iterator over the set flags and their names.

These are bitflags names in SCREAMING_SNAKE_CASE.