Struct wgpu_types::TextureFormatInfo

pub struct TextureFormatInfo {
    pub required_features: Features,
    pub sample_type: TextureSampleType,
    pub block_dimensions: (u8, u8),
    pub block_size: u8,
    pub components: u8,
    pub srgb: bool,
    pub guaranteed_format_features: TextureFormatFeatures,
}

Information about a texture format.

Fields
required_features: Features

Features required (if any) to use the texture.

sample_type: TextureSampleType

Type of sampling that is valid for the texture.

block_dimensions: (u8, u8)

Dimension of a “block” of texels. This is always (1, 1) on uncompressed textures.

block_size: u8

Size in bytes of a “block” of texels. This is the size per pixel on uncompressed textures.

components: u8

Count of components in the texture. This determines which components there will be actual data in the shader for.

srgb: bool

Format will have colors be converted from srgb to linear on read and from linear to srgb on write.

guaranteed_format_features: TextureFormatFeatures

Format features guaranteed by the WebGPU spec. Additional features are available if Features::TEXTURE_ADAPTER_SPECIFIC_FORMAT_FEATURES is enabled.

Implementations

impl TextureFormatInfo

pub fn is_compressed(&self) -> bool

Return true for compressed formats.