wgpu_types

Module surface

Structs

PresentationTimestamp

Nanosecond timestamp used by the presentation engine.

SurfaceCapabilities

Defines the capabilities of a given surface and adapter.

SurfaceConfiguration

Configures a Surface for presentation.

Enums

CompositeAlphaMode

Specifies how the alpha channel of the textures should be handled during compositing.

PresentMode

Timing and queueing with which frames are actually displayed to the user.

SurfaceStatus

Status of the received surface image.