wgpu_types

Module texture

Re-exports

pub use external_image::*;

pub use external_texture::*;

pub use format::*;

Modules

external_image 🔒

external_texture 🔒

format 🔒

Structs

ImageSubresourceRange

Subresource range within an image

SamplerDescriptor

Describes a Sampler.

TexelCopyBufferInfo

View of a buffer which can be used to copy to/from a texture.

TexelCopyBufferLayout

Layout of a texture in a buffer’s memory.

TexelCopyTextureInfo

View of a texture which can be used to copy to/from a buffer/texture.

TextureDescriptor

Describes a Texture.

TextureSelector

Specifies a particular set of subresources in a texture.

TextureTransition

A texture transition for use with CommandEncoder::transition_resources.

TextureUsages

Different ways that you can use a texture.

TextureUses

Similar to TextureUsages, but used only for CommandEncoder::transition_resources.

TextureViewDescriptor

Describes a TextureView.

Enums

AddressMode

How edges should be handled in texture addressing.

FilterMode

Texel mixing mode when sampling between texels.

MipmapFilterMode

Texel mixing mode when sampling between texels.

SamplerBorderColor

Color variation to use when sampler addressing mode is AddressMode::ClampToBorder

StorageTextureAccess

Specific type of a sample in a texture binding.

TextureAspect

Selects a subset of the data a Texture holds.

TextureDataOrder

Order in which texture data is laid out in memory.

TextureDimension

Dimensionality of a texture.

TextureSampleType

Specific type of a sample in a texture binding.

TextureViewDimension

Dimensions of a particular texture view.